Added primary directory grouping to test file list
The test_directories array in bootstrap.php gained a key for each directory. This key is used to group files and sub-directories with the given primary test directory. This keeps projects together and makes it easier to select an entire project.

The Javascript for FileSelector, as well as the FileList ajax handler were also updated. The Home controller needed a minor tweak since it expected numeric keys in the test_directories array